Use unsigned long long so numbers dont get truncated

+++ inc/internals.h
@@ -6,7 +6,7 @@
 /*   By: franmart <franmart@student.42malaga.com>   +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/09/27 18:34:25 by franmart          #+#    #+#             */
-/*   Updated: 2024/09/27 19:33:51 by franmart         ###   ########.fr       */
+/*   Updated: 2024/10/13 09:41:21 by franmart         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -28,7 +28,7 @@ typedef struct s_buffer
 int				ft_check_arg(char flag, va_list args, t_buffer *buf);
 int				ft_parse_char(char c, t_buffer *buf);
 int				ft_parse_str(char *str, t_buffer *buf);
-int				ft_parse_nbr(long n, char *base, char flag, t_buffer *buf);
+int				ft_parse_nbr(unsigned long long n, char *base, char flag, t_buffer *buf);
 unsigned int	ft_len(char *str);
 void			ft_putchar_buf(char c, t_buffer *buf);
 void			ft_putstr_buf(char *s, t_buffer *buf);
